requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-session-timeout==0.1.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0
Pillow==8.4.0
protobuf==3.19.1
six==1.16.0
sqlparse==0.4.2
